refactor: rename function name
