+++ feature/settings/src/main/java/dev/anilbeesetti/nextplayer/settings/screens/PlayerPreferencesScreen.kt
@@ -1,12 +1,8 @@
 package dev.anilbeesetti.nextplayer.settings.screens
 
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.selection.selectableGroup
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.ArrowBack
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -25,12 +21,10 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import dev.anilbeesetti.nextplayer.core.datastore.DoubleTapGesture
 import dev.anilbeesetti.nextplayer.core.datastore.Resume
-import dev.anilbeesetti.nextplayer.core.ui.components.CancelButton
-import dev.anilbeesetti.nextplayer.core.ui.components.NextDialog
-import dev.anilbeesetti.nextplayer.core.ui.components.NextDialogDefaults
 import dev.anilbeesetti.nextplayer.core.ui.components.RadioTextButton
 import dev.anilbeesetti.nextplayer.feature.settings.R
 import dev.anilbeesetti.nextplayer.settings.composables.ClickablePreferenceItem
+import dev.anilbeesetti.nextplayer.settings.composables.OptionsDialog
 import dev.anilbeesetti.nextplayer.settings.composables.PreferenceSwitch
 import dev.anilbeesetti.nextplayer.settings.composables.PreferenceSwitchWithDivider
 
@@ -112,7 +106,7 @@ fun PlayerPreferencesScreen(
                                 text = it.value,
                                 selected = (it == preferences.resume),
                                 onClick = {
-                                    viewModel.updateResume(it)
+                                    viewModel.updatePlaybackResume(it)
                                     viewModel.onEvent(PlayerPreferencesEvent.ShowDialog(Dialog.None))
                                 }
                             )
@@ -142,29 +136,4 @@ fun PlayerPreferencesScreen(
             }
         }
     }
-}
-
-@Composable
-fun OptionsDialog(
-    text: String,
-    onDismissClick: () -> Unit,
-    options: @Composable ColumnScope.() -> Unit
-) {
-    NextDialog(
-        onDismissRequest = onDismissClick,
-        title = {
-            Text(
-                text = text
-            )
-        },
-        content = {
-            Column(
-                modifier = Modifier.selectableGroup()
-            ) {
-                options()
-            }
-        },
-        dismissButton = { CancelButton(onClick = onDismissClick) },
-        confirmButton = null
-    )
 }
